add wchar test
